Separate Gradle plugin and formatting by packages
