+++ tools/javadoc-style/src/main/java/io/spine/tools/javadoc/style/gradle/JavadocStyleTaskName.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.javadoc.style;
+package io.spine.tools.javadoc.style.gradle;
 
 import io.spine.annotation.Internal;
 import io.spine.tools.gradle.TaskName;
