+++ tools/javadoc-style/src/main/resources/META-INF/gradle-plugins/io.spine.javadoc-style.properties
@@ -24,4 +24,4 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #
 
-implementation-class=io.spine.tools.javadoc.style.JavadocStylePlugin
+implementation-class=io.spine.tools.javadoc.style.gradle.JavadocStylePlugin
