+++ tools/javadoc-style/src/test/java/io/spine/tools/javadoc/style/formatting/PreTagFormattingTest.java
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.javadoc.style;
+package io.spine.tools.javadoc.style.formatting;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
