+++ tools/javadoc-style/src/test/java/io/spine/tools/javadoc/style/gradle/JavadocStylePluginFormattingTest.java
@@ -24,14 +24,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.tools.javadoc.style;
+package io.spine.tools.javadoc.style.gradle;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import io.spine.tools.gradle.testing.GradleProject;
-import org.gradle.api.Project;
-import org.gradle.testfixtures.ProjectBuilder;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -43,15 +40,11 @@ import java.nio.file.Path;
 import java.util.List;
 
 import static com.google.common.truth.Truth.assertThat;
-import static io.spine.tools.gradle.JavaTaskName.compileJava;
-import static io.spine.tools.gradle.JavaTaskName.compileTestJava;
-import static io.spine.tools.gradle.ProtobufTaskName.generateProto;
-import static io.spine.tools.gradle.ProtobufTaskName.generateTestProto;
 import static io.spine.tools.gradle.TaskDependencies.dependsOn;
-import static io.spine.tools.javadoc.style.BacktickFormatting.BACKTICK;
-import static io.spine.tools.javadoc.style.JavadocStyleTaskName.formatProtoDoc;
-import static io.spine.tools.javadoc.style.PreTagFormatting.CLOSING_PRE;
-import static io.spine.tools.javadoc.style.PreTagFormatting.OPENING_PRE;
+import static io.spine.tools.javadoc.style.formatting.BacktickFormatting.BACKTICK;
+import static io.spine.tools.javadoc.style.gradle.JavadocStyleTaskName.formatProtoDoc;
+import static io.spine.tools.javadoc.style.formatting.PreTagFormatting.CLOSING_PRE;
+import static io.spine.tools.javadoc.style.formatting.PreTagFormatting.OPENING_PRE;
 import static java.lang.System.lineSeparator;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
